// frontend-admin/src/api/generated/model/backupRunResponseDtoStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export type BackupRunResponseDtoStatus = typeof BackupRunResponseDtoStatus[keyof typeof BackupRunResponseDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BackupRunResponseDtoStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
} as const;