// frontend-admin/src/api/generated/model/backupVerificationResponseDtoStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export type BackupVerificationResponseDtoStatus = typeof BackupVerificationResponseDtoStatus[keyof typeof BackupVerificationResponseDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BackupVerificationResponseDtoStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
} as const;